Allow for random release within a vertical range
In case the initial release P0 is of the form:

P0=[172.6,.-45.6,[-3.0,-6.0]]
Then release at fixed X,Y but randomly within the range -3,-6 in the
vertical.
It could be extended to allow for range of X,Y for release along a line.
There may be a more elegant way of doing this ....